BasicFilterProto: use float for sample_rate
We generally assume that the sample rate is specified as a float, it
being specified as double here broke the (initial) settingsUpdate when
dynamically instantiating the block.

+++ blocks/filter/include/gnuradio-4.0/filter/time_domain_filter.hpp
@@ -149,9 +149,9 @@ with selectable filter type (low-pass, high-pass, band-pass, band-stop), and sup
     Annotated<FilterType, "filter_type", Doc<"Filter type ('FIR' or 'IIR')">, Visible>                                                         filter_type     = FilterType::IIR;
     Annotated<filter::Type, "filter_response", Doc<"Filter response ('LOWPASS', 'HIGHPASS', 'BANDPASS', 'BANDSTOP')">, Visible>                filter_response = filter::Type::LOWPASS;
     Annotated<gr::Size_t, "filter_order", Doc<"Filter order">>                                                                                 filter_order{3};
-    Annotated<double, "f_low", Doc<"Low cutoff frequency in Hz">, Visible>                                                                     f_low{0.1};
-    Annotated<double, "f_high", Doc<"High cutoff frequency in Hz (only for BANDPASS/BANDSTOP)">, Visible>                                      f_high{0.2};
-    Annotated<double, "sample rate", Doc<"Sample rate in Hz">, Visible>                                                                        sample_rate{1.0};
+    Annotated<float, "f_low", Doc<"Low cutoff frequency in Hz">, Visible>                                                                      f_low{0.1f};
+    Annotated<float, "f_high", Doc<"High cutoff frequency in Hz (only for BANDPASS/BANDSTOP)">, Visible>                                       f_high{0.2f};
+    Annotated<float, "sample rate", Doc<"Sample rate in Hz">, Visible>                                                                         sample_rate{1.0f};
     Annotated<gr::Size_t, "decimation factor", Doc<"1: none, i.e. preserving the relationship: N_out = N_in/decimate">>                        decimate{1U};
     Annotated<filter::iir::Design, "iir_design_method", Doc<"IIR Filter design method ('BUTTERWORTH', 'BESSEL', 'CHEBYSHEV1', 'CHEBYSHEV2')">> iir_design_method = filter::iir::Design::BUTTERWORTH;
     Annotated<algorithm::window::Type, "fir_design_method", Doc<"FIR Filter design method ('None', 'Rectangular', 'Hamming', 'Hann', 'HannExp', 'Blackman', 'Nuttall', 'BlackmanHarris', 'BlackmanNuttall', 'FlatTop', 'Exponential', 'Kaiser')">> //
@@ -171,9 +171,9 @@ with selectable filter type (low-pass, high-pass, band-pass, band-stop), and sup
         // Set up filter parameters
         FilterParameters params;
         params.order = filter_order;
-        params.fLow  = f_low;
-        params.fHigh = f_high;
-        params.fs    = sample_rate;
+        params.fLow  = static_cast<double>(f_low);
+        params.fHigh = static_cast<double>(f_high);
+        params.fs    = static_cast<double>(sample_rate);
 
         if (filter_type == FilterType::FIR) { // design FIR filter
             _filter = FilterImpl(fir::designFilter<ValueType>(filter_response, params, fir_design_method));
